support range of version numbers for scala dependencies in pom.xml
Created new property scala.version.range in build.xml and set it to
[2.10.0, 2.10.99].  New property gets used for setting <version>
value of scala-compiler and scala-library dependencies in pom.

Tried to use range [2.10.0, 2.11), but that caused scalatest to get
compiled with 2.11.  Suspect due to a bug in maven ant tools.

Tried to just change scala.version to new range value, but that caused
osgi to fail.

Deployed this commit as SNAP36.